pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "SEO Analyzer for web-projects"
authors = [{ name = "DomoSpb" }]
requires-python = "~=3.10"
readme = "README.md"
license = "MIT"
dependencies = [
    "flask>=3.0.3,<4",
    "flake8>=7.1.0,<8",
    "gunicorn>=22.0.0,<23",
    "black>=24.8.0,<25",
    "python-dotenv>=1.0.1,<2",
    "psycopg2-binary>=2.9.9,<3",
    "validators>=0.33.0,<1",
    "requests>=2.31.0,<3",
    "beautifulsoup4>=4.12.0,<5"
]

[tool.hatch.build.targets.sdist]
include = ["page_analyzer"]

[tool.hatch.build.targets.wheel]
include = ["page_analyzer"]

[tool.pytest]
test = "pytest --cov=page_analyzer"
test-coverage = "pytest --cov=page_analyzer --cov-report xml"